traitements/shared/Modele/analyseur_loi/requirements.txt
boto3>=1.26.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
beautifulsoup4>=4.11.0
transformers>=4.21.0
spacy>=3.4.0
jsonschema>=4.17.0
python-dotenv>=0.19.0
retrying>=1.3.3
streamlit>=1.36.0
plotly>=5.20.0
